backend/requirements.txt
alembic==1.14.1
asgiref==3.7.2
attrs==23.1.0
Authlib==1.4.0
bcrypt==4.2.1
beautifulsoup4==4.12.2
black==23.12.1
blinker==1.9.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
coverage==7.3.4
cryptography==44.0.0
cycler==0.12.1
distlib==0.3.7
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
dnspython==2.6.1
filelock==3.13.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mypy-extensions==1.0.0
numpy==2.2.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.9.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
selenium==4.16.0
setuptools==69.0.2
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
threadpoolctl==3.5.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.1.0
virtualenv==20.25.0
webdriver-manager==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
